:cherries:Implement more lighting maps. Enable light settings in the INI. Improve terrain shader to handle different type of texture. Turned on compiler warning flags and optimise the program based on compiler warnings. Fix incorrect terrain normal direction.

+++ SuperTerrain+/Core/STPTextureDatabase.cpp
@@ -141,7 +141,7 @@ public:
 	*/
 	STPSmartStmt createStmt(const string_view& sql, unsigned int flag = 0u) {
 		sqlite3_stmt* newStmt;
-		STPsqliteCheckErr(sqlite3_prepare_v3(this->SQL, sql.data(), sql.size(), flag, &newStmt, nullptr));
+		STPsqliteCheckErr(sqlite3_prepare_v3(this->SQL, sql.data(), static_cast<int>(sql.size()), flag, &newStmt, nullptr));
 		return STPSmartStmt(newStmt);
 	}
